third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/iframes-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElementImage does not draw cross-origin iframes</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='../support/getPixel.js'></script>
  <script src="/wpt_internal/resources/canvas-draw-element/waitForCanvasPaint.js"></script>
  <style>
  #child {
    width: 100px;
    height: 200px;
    background: blue;
  }
  #sameOrigin {
    position: absolute;
    left: 0px;
    top: 0px;
  }
  #crossOrigin {
    position: absolute;
    left: 0px;
    top: 100px;
  }
  </style>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <div id=child>
      <iframe id=sameOrigin width=100 height=100 src="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/resources/iframe-subframe-green.html"></iframe>
      <iframe id=crossOrigin width=100 height=100 src="https://{{hosts[alt][www]}}:{{ports[https][0]}}/wpt_internal/html/canvas/drawElementImage/resources/iframe-subframe-red.html"></iframe>
    </div>
  </canvas>

  <script>
    let sPendingIframesToLoad = 2;

    function pendingIframeLoaded() {
      if (!--sPendingIframesToLoad) {
        promise_test(async function(t) {
          await waitForCanvasPaint(canvas);
          const context = canvas.getContext("2d");
          context.drawElementImage(child, 0, 0);

          // Fetch all pixel data once to avoid multiple slow readbacks.
          const imgData = context.getImageData(0, 0, canvas.width, canvas.height);

          let pixel = getPixel(imgData, 50, 50);
          assert_array_equals(pixel, [0, 128, 0, 255], "Same origin should draw");

          pixel = getPixel(imgData, 50, 150);
          assert_array_equals(pixel, [0, 0, 255, 255], "Cross origin should not draw");
        });
      }
    }

    sameOrigin.onload = pendingIframeLoaded;
    crossOrigin.onload = pendingIframeLoaded;
  </script>
</body>
</html>
